refactor(folder): rename folder for auth forms (modals)
