+++ frontend/src/authModalComponents/FormImgComponent/FormImgComponent.module.css
