+++ frontend/src/authModalComponents/LoginForm/LoginForm.module.css
@@ -1,0 +1,58 @@
+.container {
+  display: flex;
+}
+
+.formWrapper {
+  padding: 12px 20px;
+}
+
+.headerWrapper {
+  display: flex;
+  margin-bottom: 24px;
+}
+
+.headerWrapLogin,
+.headerWrapRegister {
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  width: 180px;
+  padding: 8px 12px;
+}
+
+.headerWrapLogin::after {
+  position: absolute;
+  width: 184px;
+  height: 1.5px;
+  left: -2px;
+  bottom: -1px;
+  background-color: var(--primary-yellow);
+  content: '';
+}
+
+.headerBtnActive {
+  text-align: center;
+  font-size: var(--font-size-small);
+  font-weight: var(--font-weight-bold);
+  border: none;
+  background-color: transparent;
+  color: var(--default-black);
+}
+
+.headerBtn {
+  text-align: center;
+  font-size: var(--font-size-small);
+  font-weight: var(--font-weight-bold);
+  border: none;
+  background-color: transparent;
+  color: var(--grey-darker);
+}
+
+.privacyText {
+  margin-bottom: 16px;
+  font-size: var(--font-size-extra-tiny);
+  font-weight: var(--font-weight-normal);
+  color: var(--default-black);
+}
