--- frontend/src/formModalComponents/LoginForm/LoginForm.module.css
@@ -1,240 +1,0 @@
-.container {
-  display: flex;
-}
-
-.formWrapper {
-  padding: 12px 20px;
-}
-
-.headerWrapper {
-  display: flex;
-  margin-bottom: 24px;
-}
-
-.headerWrapLogin,
-.headerWrapRegister {
-  margin: 0 auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-  width: 180px;
-  padding: 8px 12px;
-}
-
-.headerWrapLogin::after {
-  position: absolute;
-  width: 184px;
-  height: 1.5px;
-  left: -2px;
-  bottom: -1px;
-  background-color: var(--primary-yellow);
-  content: '';
-}
-
-.headerBtnActive {
-  text-align: center;
-  font-size: var(--font-size-small);
-  font-weight: var(--font-weight-bold);
-  border: none;
-  background-color: transparent;
-  color: var(--default-black);
-}
-
-.headerBtn {
-  text-align: center;
-  font-size: var(--font-size-small);
-  font-weight: var(--font-weight-bold);
-  border: none;
-  background-color: transparent;
-  color: var(--grey-darker);
-}
-
-/*=========START PWD INPUT=========*/
-.pwdInputWrap {
-  margin-bottom: 8px;
-}
-
-.pwdLabelWrap {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.inputLabel {
-  margin-right: 4px;
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-bold);
-  color: var(--default-black);
-}
-
-.requiredSymb {
-  font-size: var(--font-size-extra-small);
-  font-weight: var(--font-weight-bold);
-  color: var(--default-black);
-}
-
-.resetPwd {
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-normal);
-  border: none;
-  background-color: transparent;
-  color: var(--default-black);
-}
-
-.pwdInputBox {
-  position: relative;
-}
-
-.formInput {
-  width: 368px;
-  margin-top: 4px;
-  height: 44px;
-  padding: 0 12px;
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-normal);
-  border-radius: var(--border-radius);
-  border: var(--border-width) var(--border-style) var(--grey-dark);
-  color: var(--default-black);
-  background-color: transparent;
-  appearance: none;
-}
-
-.filled {
-  border-color: var(--default-black);
-}
-
-.formInput:hover {
-  border-color: var(--primary-yellow);
-}
-
-.formInput:focus {
-  outline: var(--border-width) var(--border-style) transparent;
-  border-color: var(--primary-yellow);
-}
-
-.formInput::placeholder {
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-normal);
-  color: var(--grey-dark);
-}
-
-.formInputError {
-  border: var(--border-width) var(--border-style) var(--error-red);
-}
-
-.fiEyeOff {
-  position: absolute;
-  right: 12px;
-  top: -20%;
-  transform: translateY(100%);
-  width: var(--icon-size-normal);
-  height: var(--icon-size-normal);
-  cursor: pointer;
-  stroke: var(--grey-dark);
-}
-
-.fiEye {
-  position: absolute;
-  right: 12px;
-  top: -20%;
-  transform: translateY(100%);
-  width: var(--icon-size-normal);
-  height: var(--icon-size-normal);
-  cursor: pointer;
-  stroke: var(--grey-darkest);
-}
-
-.iconError {
-  stroke: var(--error-red);
-}
-
-.additionalInfo {
-  margin-top: 4px;
-  font-size: var(--font-size-extra-tiny);
-  font-weight: var(--font-weight-normal);
-  color: var(--error-red);
-}
-/*=========END INPUT PWD=========*/
-
-/*=========START CHECKBOX=========*/
-.visuallyHidden {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  padding: 0;
-  white-space: nowrap;
-  overflow: hidden;
-  border: 0;
-  clip-path: inset(100%);
-  clip: rect(0 0 0 0);
-}
-
-.rememberCheckbox {
-  display: flex;
-  align-items: center;
-  justify-content: start;
-  margin-bottom: 24px;
-}
-
-.rememberCheckboxLabel {
-  display: flex;
-  align-items: center;
-}
-
-.userRememberComponentIcon {
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  width: var(--checkbox-size);
-  height: var(--checkbox-size);
-  background-color: transparent;
-  border: var(--border-width) var(--border-style) var(--grey-dark);
-  border-radius: var(--border-radius-small);
-  margin-right: 9px;
-}
-
-.checkmarkIcon {
-  display: none;
-  color: var(--default-white);
-}
-
-.visuallyHidden:checked + .rememberCheckboxLabel .checkmarkIcon {
-  display: inline;
-}
-
-input[type='checkbox']:checked
-  + .rememberCheckboxLabel
-  > .userRememberComponentIcon {
-  background-color: var(--primary-yellow);
-  border-color: var(--primary-yellow);
-}
-
-input[type='checkbox']:hover
-  + .rememberCheckboxLabel
-  > .userRememberComponentIcon {
-  outline: var(--border-width) var(--border-style) transparent;
-  border: var(--border-width) var(--border-style) var(--primary-yellow);
-}
-
-input[type='checkbox']:focus
-  + .rememberCheckboxLabel
-  > .userRememberComponentIcon {
-  outline: var(--border-width) var(--border-style) transparent;
-  border: var(--border-width) var(--border-style) var(--primary-yellow);
-}
-
-.rememberText {
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-normal);
-  color: var(--default-black);
-}
-
-.privacyText {
-  margin-bottom: 16px;
-  font-size: var(--font-size-extra-tiny);
-  font-weight: var(--font-weight-normal);
-  color: var(--default-black);
-}
-/*=========END CHECKBOX=========*/
