--- frontend/src/formModalComponents/PasswordField/PasswordField.module.css
@@ -1,98 +1,0 @@
-.pwdFieldWrap {
-  margin-bottom: 8px;
-}
-
-.pwdFieldLabel {
-  margin-right: 4px;
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-bold);
-  color: var(--default-black);
-}
-
-.requiredSymb {
-  font-size: var(--font-size-extra-small);
-  font-weight: var(--font-weight-bold);
-  color: var(--default-black);
-}
-
-.resetPwd {
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-normal);
-  border: none;
-  background-color: transparent;
-  color: var(--default-black);
-}
-
-.wrap {
-  position: relative;
-}
-
-.pwdField {
-	width: 368px; 
-  margin-top: 4px;
-  height: 44px;
-  padding: 0 12px;
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-normal);
-  border-radius: var(--border-radius);
-  border: var(--border-width) var(--border-style) var(--grey-dark);
-  color: var(--default-black);
-  background-color: transparent;
-  appearance: none;
-}
-
-.pwdFieldFilled {
-  border-color: var(--default-black);
-}
-
-.pwdFieldt:hover {
-  border-color: var(--primary-yellow);
-}
-
-.pwdField:focus {
-  outline: var(--border-width) var(--border-style) transparent;
-  border-color: var(--primary-yellow);
-}
-
-.pwdField::placeholder {
-  font-size: var(--font-size-tiny);
-  font-weight: var(--font-weight-normal);
-  color: var(--grey-dark);
-}
-
-.pwdFieldError {
-  border: var(--border-width) var(--border-style) var(--error-red);
-}
-
-.fiEyeOff {
-  position: absolute;
-  right: 12px;
-  top: -20%;
-  transform: translateY(100%);
-  width: var(--icon-size-normal);
-  height: var(--icon-size-normal);
-  cursor: pointer;
-  stroke: var(--grey-dark);
-}
-
-.fiEye {
-  position: absolute;
-  right: 12px;
-  top: -20%;
-  transform: translateY(100%);
-  width: var(--icon-size-normal);
-  height: var(--icon-size-normal);
-  cursor: pointer;
-  stroke: var(--grey-darkest);
-}
-
-.iconError {
-  stroke: var(--error-red);
-}
-
-.additionalInfo {
-  margin-bottom: 8px;
-  font-size: var(--font-size-extra-tiny);
-  font-weight: var(--font-weight-normal);
-  color: var(--grey-darker);
-}
