--- frontend/src/formModalComponents/PwdStrengthLength/PwdStrengthLength.module.css
@@ -1,24 +1,0 @@
-.pwdStrengthContainer {
-  display: flex;
-  justify-content: start;
-  align-items: center;
-  margin: 4px 0px;
-}
-
-.strengthLengthWrap {
-  display: flex;
-}
-
-.pwdStrengthLength {
-  display: flex;
-  height: 4px;
-  border-radius: 2px;
-  width: 56px;
-  margin-right: 4px;
-}
-
-.strengthLabel {
-  font-size: var(--font-size-extra-tiny);
-  font-weight: var(--font-weight-normal);
-  color: var(--default-black);
-}
